DPRO-895: Disable auditing
Introduces a global flag that causes all audit service methods to be
skipped, with SqlService throwing an exception if any service attempts to
write or read to the Audit table. Audit-related code in unit tests is
conditionally skipped.

See comments on BaseRepoService.AUDITING_ENABLED: this is temporary until a
future release.